+++ CHANGELOG.md
@@ -8,5 +8,6 @@ All notable changes to the "Chat Program" project will be documented in this fil
 ### Added
 
 - Initial project structure.
+- Initial socket, express, ejs as template and insert a chat application ejs temp
 
 ---